+++ home/src/components/Home.tsx
@@ -7,10 +7,8 @@ import Articles from "./Articles";
 import { useEffect } from "react";
 import { applyTheme } from "../themes/utils";
 import baseTheme from "../themes/base";
-import darkTheme from "../themes/dark";
 
 import BioCard from "./BioCard";
-import SkewedCard from "./SkewedCard";
 
 export default function Home() {
   useEffect(() => {
@@ -18,15 +16,17 @@ export default function Home() {
   }, []);
 
   return (
-    <div>
+    <div
+      id="mainContainer"
+      className="flex items-center justify-center h-screen bg-fixed bg-center custom-img"
+    >
       <BioCard />
 
-      <div className="bg-primary-light py-10">
+      <div className="bg-primary-light backgroundPattern">
         <Articles />
       </div>
 
       <StickyFooter />
-
     </div>
   );
 }
