+++ home/src/themes/darkOld.tsx
@@ -1,0 +1,21 @@
+import { createTheme } from "./utils";
+
+const darkTheme = createTheme({
+  desc: "dark",
+  primary: "#373737",
+  primaryLight: "#626262",
+  primaryDark: "#121212",
+  secondary: "#4DB6AC",
+  secondaryLight: "#82E9DE",
+  secondaryDark: "#00867D",
+  tertiary: "#CCCCCC",
+  tertiaryLight: "#82E9DE",
+  tertiaryDark: "#00867D",  
+  textOnPrimary: "#717171",
+  textOnSecondary: "#588487",
+  accentPrimary: "#588487",
+  accentSecondary: "#717171",
+  textBase: "#FFFFFF",
+});
+
+export default darkTheme;
