--- home/src/themes/git.tsx
@@ -1,24 +1,0 @@
-import { createTheme } from "./utils";
-
-const darkTheme = createTheme({
-  primary: "#0D1117",
-  primaryLight: "#161B22",
-  primaryDark: "#010409",
-  
-  secondary: "#1F6FEB",
-  secondaryLight: "#79C0FF",
-  secondaryDark: "#1951A8",
-  
-  tertiary: "#373737",
-  tertiaryLight: "#79C0FF",
-  tertiaryDark: "#1951A8",
-  
-  textOnPrimary: "#CCCCCC",
-  textOnSecondary: "#F0F6D8",
-  
-  accentPrimary: "#238636",
-  accentSecondary: "#E9AA1E",
-  textBase: "#F0F6D8",
-});
-
-export default darkTheme;
